+++ get_input.h
@@ -3,7 +3,7 @@
 #ifndef GET_INPUT_H
 #define GET_INPUT_H
 
-int get_input (float *, float *, unsigned char *);
+int get_input (double *, double *, unsigned char *);
 
 
 #endif
